2024 05 17 18:25 midsave

+++ store_resource_server/src/main/java/ru/gb/group5984/service/api/CharacterApiServiceImpl.java
@@ -1,5 +1,7 @@
 package ru.gb.group5984.service.api;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.java.Log;
@@ -8,12 +10,15 @@ import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import ru.gb.group5984.aspect.TrackUserAction;
+import ru.gb.group5984.auth.AuthenticationService;
 import ru.gb.group5984.configuration.BasicConfig;
 import ru.gb.group5984.model.characters.CharacterResult;
 import ru.gb.group5984.model.characters.Characters;
+import ru.gb.group5984.model.messeges.Message;
 import ru.gb.group5984.model.transactions.Transaction;
 import ru.gb.group5984.service.db.ServerDbService;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -26,6 +31,7 @@ import java.util.List;
 public class CharacterApiServiceImpl  implements CharacterApiService{
     private final ServerDbService serverDbService;
     private final BasicConfig basicConfig;
+    private final AuthenticationService authenticationService;
 
     @Autowired
     private RestTemplate restTemplate;
@@ -80,16 +86,38 @@ public class CharacterApiServiceImpl  implements CharacterApiService{
     @Override
     @Transactional
     @TrackUserAction
-    public void basketPay() {
-        Double totalAmount = serverDbService.getTotalPriceFromBasket();
+    public Message basketPay() {
+        BigDecimal totalAmount = serverDbService.getTotalPriceFromBasket();
         String url = basicConfig.getBANK_API() + "/transaction";
-        Transaction transaction = new Transaction();
-        transaction.setCreditAccount(2L); // Пока тест
-        transaction.setDebitAccount(1L); // Пока тест
-        transaction.setTransferAmount(totalAmount);
+        HttpMethod method = HttpMethod.POST;
+                Class<Message> responseType = Message.class;
+        var transaction = Transaction.builder()
+                .creditAccount(2L)
+                .debitAccount(1L)
+                .transferAmount(totalAmount)
+                .build();
+        String jsonTransaction = "";
+        try {
+            jsonTransaction = new ObjectMapper().writeValueAsString(transaction);
+        } catch (JsonProcessingException ignored){
+        }
+
+        HttpEntity<String> requestEntity = new HttpEntity<>(jsonTransaction, authenticationService.getHeaders());
         log.info("Тест платежа: " + transaction);
         log.info("URL: " + url);
-        restTemplate.postForEntity(url, transaction, Transaction.class);
-        serverDbService.deleteAllFromBasket();
+        Message message = new Message();
+        try {
+            ResponseEntity<Message> response = restTemplate
+                    .exchange(url, method, requestEntity, responseType);
+            message = response.getBody();
+            assert message != null;
+            if (message.getMessage().equals("OK")) {
+                serverDbService.deleteAllFromBasket();
+                message.setMessage("Оплата товара прошла успешно. Поздравляем с покупкой!");
+            }
+        } catch (RuntimeException e) {
+            message.setMessage("Ресурс банка временно недоступен: " + e.getMessage());
+        }
+        return message;
     }
 }
